+++ emacs-macros/src/func.rs
@@ -238,7 +238,7 @@ impl LispFunc {
             #define_wrapper
             fn #exporter(env: &::emacs::Env) -> ::emacs::Result<()> {
                 let prefix = ::emacs::init::lisp_path(#path);
-                ::emacs::export_functions! {
+                ::emacs::__export_functions! {
                     env, prefix, {
                         #lisp_name => (#wrapper, #min..#max, #doc),
                     }