+++ src/types/string.rs
@@ -1,4 +1,4 @@
-use std::{os, ptr, borrow::Borrow};
+use std::{os, ptr};
 
 use super::*;
 
@@ -17,9 +17,18 @@ impl FromLisp<'_> for String {
     }
 }
 
-impl<'e, 'a, T: Borrow<str> + ?Sized> IntoLisp<'e> for &'a T {
+impl<'e> IntoLisp<'e> for &str {
     fn into_lisp(self, env: &'e Env) -> Result<Value<'_>> {
-        let bytes = self.borrow().as_bytes();
+        let bytes = self.as_bytes();
+        let len = bytes.len();
+        let ptr = bytes.as_ptr();
+        raw_call_value!(env, make_string, ptr as *const os::raw::c_char, len as isize)
+    }
+}
+
+impl<'e> IntoLisp<'e> for &String {
+    fn into_lisp(self, env: &'e Env) -> Result<Value<'_>> {
+        let bytes = self.as_bytes();
         let len = bytes.len();
         let ptr = bytes.as_ptr();
         raw_call_value!(env, make_string, ptr as *const os::raw::c_char, len as isize)