+++ test-module/src/call.rs
@@ -35,3 +35,17 @@ fn list(env: &Env, n: u16) -> Result<Value> {
     }
     env.call(list, ints.as_slice())
 }
+
+#[defun]
+fn value<'e>(function: Value<'e>, arg: Value) -> Result<Value<'e>> {
+    function.call([arg])
+}
+
+//#[defun]
+//fn mapc_vec(function: Value, vector: emacs::Vector) -> Result<()> {
+//    for i in 0..vector.size()? as usize {
+//        let elem: Value = vector.get(i)?;
+//        function.call((i, elem))?;
+//    }
+//    Ok(())
+//}