+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
-- Made `env.call` a lot more flexible, fixing #15.
+- Made `env.call` a lot more flexible, fixing #15. Also added `value.call`.
 - Added `rust-wrong-type-user-ptr` to `wrong-type-argument` classification.
 - Deprecated `emacs::module_init!` and `emacs::export_functions!`.
 - Replaced`IntoLisp` implementation for `AsRef<str>` with separate implementations for `&str` and `&String`.